fix: import paths after renaming folders
